requirements.txt
Django
Django-highcharts
bs4
lxml
yara-python
yara
pefile
numpy
requests
scikit-learn
scipy
PyYAML